Level 1-3 in .json and Level change in Python
First 3 levels are now remade in .json.
The scene change in-game is now done in Python, which allows an arbitrary
next level name, found in the .json as a map property "next_level".
Currently crashes when the next level doesn't exists.
Improved build script to pop-up a file selector window.